requirements.txt
faiss==1.9.0
numpy==2.3.5
Pillow==12.0.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
